src/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
tqdm>=4.66.0,<5.0.0
tiktoken>=0.7.0,<1.0.0
qwen-agent>=0.0.34,<0.1.0
transformers>=4.46.0,<6.0.0
soundfile>=0.13.1,<1.0.0
